Slim down context
